backend/requirements.txt
celery==5.3.6
commonregex==1.5.4
Flask==2.2.5
Flask_Cors==4.0.0
langdetect==1.0.9
langid==1.1.6
lxml==4.9.3
mlxtend==0.23.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
presidio_analyzer==2.2.351
Requests==2.31.0
ruptures==1.1.8
scikit_learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0
textblob==0.17.1
textstat==0.7.3
Werkzeug==2.2.3
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
redis==5.0.1
kombu==5.3.4
amqp==5.2.0
gunicorn==21.2.0
firebase-admin==6.3.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0